Add files via upload
We take good care of seniors who can not be left alone, whose family members are no longer or live in long distances and those who are looking to make the best of their retirement experiences. We assist them with ADL ( activities of daily living) , create and share enjoyable moments, support and love them. We are great companions and compationate health care workers
Come be a part of our family where hearts connect and lauhter resonates . Majorie's House is not just a place; it is a feeling of home , safety , and love.